+++ tortoise/backends/asyncpg/executor.py
@@ -7,11 +7,12 @@ from pypika.terms import Term
 
 from tortoise import Model
 from tortoise.backends.base.executor import BaseExecutor
+from tortoise.contrib.postgres.search import SearchCriterion
 from tortoise.filters import search
 
 
-def postgres_search(field: Term, value: str):
-    pass
+def postgres_search(field: Term, value: Term):
+    return SearchCriterion(field, expr=value)
 
 
 class AsyncpgExecutor(BaseExecutor):